refactor form logic
